+++ app/server/beam/tau/lib/tau_web/router.ex
@@ -20,6 +20,8 @@ defmodule TauWeb.Router do
     get "/", PageController, :index
   end
 
+
+
   # Other scopes may use custom stacks.
   # scope "/api", TauWeb do
   #   pipe_through :api
@@ -35,25 +37,15 @@ defmodule TauWeb.Router do
   if Mix.env() in [:dev, :test] do
     import Phoenix.LiveDashboard.Router
 
-    scope "/" do
+    scope "/dev", TauWeb do
       pipe_through :browser
+
+      live "/log", LogLive, :index
       live_dashboard "/dashboard",
         metrics: TauWeb.Telemetry,
         additional_pages: [
           flame_on: FlameOn.DashboardPage
           ]
-    end
-  end
-
-  # Enables the Swoosh mailbox preview in development.
-  #
-  # Note that preview only shows emails that were sent by the same
-  # node running the Phoenix server.
-  if Mix.env() == :dev do
-    scope "/dev" do
-      pipe_through :browser
-
-      forward "/mailbox", Plug.Swoosh.MailboxPreview
     end
   end
 end