+++ app/server/ruby/lib/sonicpi/lang/core.rb
@@ -1752,7 +1752,7 @@ end"
         return [].ring if start == finish
 
         raise ArgumentError, "step size: opt for fn range should be a non-zero number" unless step_size != 0
-        
+
         step_size = step_size.abs
         res = []
         cur = start
@@ -4214,19 +4214,18 @@ puts current_sched_ahead_time # Prints 0.5"]
         # Schedule messages
         __schedule_delayed_blocks_and_messages!
         return if beats == 0
-        __system_thread_locals.set(:sonic_pi_spider_slept, true)
+
         __change_spider_beat_and_time_by_beat_delta!(beats)
 
         sat = current_sched_ahead_time
-        new_vt = __get_spider_time.to_r
+        new_vt = __get_spider_time.to_f
         now = Time.now.to_f
 
         in_time_warp = __system_thread_locals.get(:sonic_pi_spider_in_time_warp)
 
-        if (now - (sat + 0.5)) > new_vt
-
-          # raise TimingError, "Timing Exception: thread got too far behind time
+        if (now - (sat + 1)) > new_vt
           __delayed_serious_warning "Serious timing error. Too far behind time..."
+          raise TimingError, "Timing Exception: thread got too far behind time"
         elsif (now - sat) > new_vt
           __delayed_serious_warning "Timing error: can't keep up..."
         elsif now > new_vt
@@ -4235,18 +4234,15 @@ puts current_sched_ahead_time # Prints 0.5"]
           unless __thread_locals.get(:sonic_pi_mod_sound_synth_silent) || in_time_warp
             __delayed_warning "Timing warning: running slightly behind..."
           end
-        else
-          if in_time_warp
-            # Don't sleep if within a time warp
-            #
-            # However, do make sure the vt hasn't got too far ahead of the real time
-            # raise TimingError, "Timing Exception: thread got too far ahead of time" if  (new_vt - 17) > now
-          else
-            t = (new_vt - now).to_f
-            Kernel.sleep t
-          end
         end
 
+        if in_time_warp
+          # Don't physically sleep or register as slept within a time warp
+        else
+          t = (new_vt - now).to_f - 0.2
+          Kernel.sleep t if t > 0.2
+          __system_thread_locals.set(:sonic_pi_spider_slept, true)
+        end
 
         ## reset control deltas now that time has advanced
         __system_thread_locals.set_local :sonic_pi_local_control_deltas, {}